cont. refactor docker utils
