+++ bentoml/yatai/deployment/docker_utils.py
@@ -1,6 +1,11 @@
+import json
+import logging
 import subprocess
 
 from bentoml.exceptions import BentoMLException, MissingDependencyException
+
+
+logger = logging.getLogger(__name__)
 
 
 def ensure_docker_available_or_raise():
@@ -15,3 +20,30 @@ def ensure_docker_available_or_raise():
             'Docker is required for this deployment. Please visit '
             'www.docker.com for instructions'
         )
+
+
+def process_docker_api_line(payload):
+    """ Process the output from API stream, throw an Exception if there is an error """
+    # Sometimes Docker sends to "{}\n" blocks together...
+    errors = []
+    for segment in payload.decode("utf-8").strip().split("\n"):
+        line = segment.strip()
+        if line:
+            try:
+                line_payload = json.loads(line)
+            except ValueError as e:
+                logger.warning("Could not decipher payload from Docker API: %s", str(e))
+            if line_payload:
+                if "errorDetail" in line_payload:
+                    error = line_payload["errorDetail"]
+                    error_msg = 'Error running docker command: {}: {}'.format(
+                        error["code"], error['message']
+                    )
+                    logger.error(error_msg)
+                    errors.append(error_msg)
+                elif "stream" in line_payload:
+                    logger.info(line_payload['stream'])
+
+    if errors:
+        error_msg = ";".join(errors)
+        raise BentoMLException("Error running docker command: {}".format(error_msg))
