+++ bentoml/yatai/deployment/utils.py
@@ -12,42 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import json
-import logging
-
-from bentoml.exceptions import (
-    BentoMLException,
-    InvalidArgument,
-)
-
-logger = logging.getLogger(__name__)
-
-
-def process_docker_api_line(payload):
-    """ Process the output from API stream, throw an Exception if there is an error """
-    # Sometimes Docker sends to "{}\n" blocks together...
-    errors = []
-    for segment in payload.decode("utf-8").strip().split("\n"):
-        line = segment.strip()
-        if line:
-            try:
-                line_payload = json.loads(line)
-            except ValueError as e:
-                logger.warning("Could not decipher payload from Docker API: %s", str(e))
-            if line_payload:
-                if "errorDetail" in line_payload:
-                    error = line_payload["errorDetail"]
-                    error_msg = 'Error running docker command: {}: {}'.format(
-                        error["code"], error['message']
-                    )
-                    logger.error(error_msg)
-                    errors.append(error_msg)
-                elif "stream" in line_payload:
-                    logger.info(line_payload['stream'])
-
-    if errors:
-        error_msg = ";".join(errors)
-        raise BentoMLException("Error running docker command: {}".format(error_msg))
+from bentoml.exceptions import InvalidArgument
 
 
 def raise_if_api_names_not_found_in_bento_service_metadata(metadata, api_names):
